fixed stickiness, implemented cross-user panel creation sync, selective io emit, better id

+++ res/static/main/panel.js
@@ -1,22 +1,33 @@
 class Panel {
 
-  constructor(type="launchPanel", id='0') {
+  constructor(type="launchPanel", id=getDefaultID()) {
     this.iden = {};
     this.iden.type = type;
     this.iden.id = id;
   }
 
+  //temporarily prevents iframes from messing with dragging/resizing
+  //small todo - better way?
+  documentOverlay(bool) {
+    if (bool) {
+      document.getElementById("overlay").style.display = "block";
+    } else {
+      document.getElementById("overlay").style.display = "none";
+    }
+  }
+
   //todo fix repetitiveness. Is there a way to combine these that doesn't make things worse?
   //todo is setting them directly via .onmousewhatever good practice? find out, fix
-  //fix jumping in resize/dragging (simple math issue, calculating offsets wrong currently)
   //todo don't alert on tiny moves
   resizing(clickLeft, clickTop) {
     log(["start resize drag from " + clickLeft + ", " + clickTop + " for panel: ", this]);
+    this.documentOverlay(true);
     document.onmouseup = () => {
       log(["end resize drag on panel: ", this]);
       document.onmouseup = null;
       document.onmousemove = null;
       this.alertResize();
+      this.documentOverlay(false);
     };
     document.onmousemove = (event) => {
       var height = event.clientY - this.getTop();
@@ -29,11 +40,13 @@ class Panel {
     log(["start relocation drag from " + clickLeft + ", " + clickTop + " for panel: ", this]);
     var offsetTop = this.getTop() - clickTop;
     var offsetLeft = this.getLeft() - clickLeft;
+    this.documentOverlay(true);
     document.onmouseup = () => {
       log(["end relocation drag on panel: ", this]);
       document.onmouseup = null;
       document.onmousemove = null;
       this.alertMove();
+      this.documentOverlay(false);
     };
     document.onmousemove = (event) => {
       var top = offsetTop + event.clientY;
@@ -53,7 +66,7 @@ class Panel {
   }
 
   deleteSignal() {
-    this.passm({appl:[this.iden.id], action: 'removePanel'});
+    this.buildMessageAndSend('removePanel');
     this.delete();
   }
 
@@ -62,11 +75,11 @@ class Panel {
     this.element.remove();
   }
 
-  createNew(type="launchPanel", id='0') {
-    addPanel(new Panel(type, id));
+  createNew(type="launchPanel", signal) {
+    addPanel(new Panel(type), signal);
   }
 
-  initElement(parent) {
+  initElement(parent, signal=false) {
     var template = document.getElementById("panelTemplate");
     this.element = document.importNode(template.content, true).firstElementChild;
     this.element.children[1].setAttribute('src', "/main/panels/" + this.iden.type);
@@ -74,7 +87,7 @@ class Panel {
     //so that element-return DOM requests can access panel-specific//
     this.element.children[1].onload = () => {
       log(["iframe loaded for panel: ", this]);
-      this.passm({action: 'initPanel'});
+      this.buildMessageAndSend('initPanel', [this.iden], {signal: signal});
     };
     parent.appendChild(this.element);
   }
@@ -84,50 +97,65 @@ class Panel {
   }
 
   alertMove() {
-    this.passm({appl:[this.iden.id], action: 'movePanel', content: {loc: {top: this.getTop(), left: this.getLeft()}}});
+    this.buildMessageAndSend('movePanel', [this.iden], {loc: {top: this.getTop(), left: this.getLeft()}});
   }
 
   alertResize() {
-    this.passm({appl:[this.iden.id], action: 'resizePanel', content: {loc: {width: this.getWidth(), height: this.getHeight()}}});
+    this.buildMessageAndSend('resizePanel', [this.iden], {loc: {width: this.getWidth(), height: this.getHeight()}});
   }
 
-  passm(msg) {
+  //action: string. appl: who it applies to, content, other
+  buildMessageAndSend(action, appl=[this.iden], content, other) {
+    var msg = typeof other === 'object' ? other : {};
+    if (typeof content !== 'undefined') msg.content = content;
+    msg.appl = appl;
+    msg.action = action;
     msg.from = this.iden;
-    sendm(msg);
+    sendMessageToServer(msg);
   }
 
-  putm(msg) {
-    if (msg && typeof(msg.action) !== 'undefined') {
-      if (typeof(msg.appl) !== 'undefined' && msg.appl.includes(this.iden.id)) {
-        switch(msg.action) {
-          case 'removePanel':
-            this.delete();
-            break;
-          case 'movePanel':
-            this.updPos(msg.content.loc.top, msg.content.loc.left);
-            break;
-          case 'resizePanel':
-            this.updSize(msg.content.loc.width, msg.content.loc.height);
-            break;
-          default: //otherwise passes it on to the child to figure out
-            if (this.child && typeof(this.child.putm) === "function") {
-              this.child.putm(msg);
-            }
-            break;
-          }
-          //TODO - at the moment, if you spawn another window of the same type
-          //before the other one has received the panelInit response, they
-          //might get confused and both end up with the same ID
-          //low priority: difficult to cause, few downsides (you just delete the dup)
-        } else if (msg.action == 'initPanel' && msg.content.type == this.iden.type
-                          && (this.iden.id == '0' || msg.content.id == this.iden.id)) {
-          this.updID(msg.content.id);
-          this.updPos(msg.content.loc.top, msg.content.loc.left);
-          this.updSize(msg.content.loc.width, msg.content.loc.height);
-          this.element.children[1].contentWindow.init(this);
+  //some way to get array.includes to compare like this?
+  includesThis(applArray) {
+    if (!Array.isArray(applArray)) return false;
+    for (const identifier of applArray) {
+      if (identifier.id == this.iden.id && identifier.type == this.iden.type) {
+        return true;
       }
     }
+    return false;
   }
+
+  firstTimeUpdate(content) {
+    log(["running first time seupt for panel: ", this]);
+    this.updID(content.id);
+    this.updPos(content.loc.top, content.loc.left);
+    this.updSize(content.loc.width, content.loc.height);
+    this.element.children[1].contentWindow.init(this);
+  }
+
+  passMessageOn(msg) {
+    if (msg && typeof(msg.action) !== 'undefined' && this.includesThis(msg.appl)) {
+      switch(msg.action) {
+        case 'initPanel':
+          this.firstTimeUpdate(msg.content);
+          break;
+        case 'removePanel':
+          this.delete();
+          break;
+        case 'movePanel':
+          this.updPos(msg.content.loc.top, msg.content.loc.left);
+          break;
+        case 'resizePanel':
+          this.updSize(msg.content.loc.width, msg.content.loc.height);
+          break;
+        default: //otherwise passes it on to the child to figure out //todo function check on server init instead?
+          if (this.child && typeof(this.child.passMessageOn) === "function") {
+            this.child.passMessageOn(msg);
+          }
+          break;
+        }
+      }
+    }
 
   //############GET/SET##################
 
@@ -151,10 +179,15 @@ class Panel {
     this.iden.id = id;
   }
 
+  getIdentification() {
+    return this.iden;
+  }
+
   getID() {
     return this.iden.id;
   }
 
+//we could collapse this into one function (type) (this.element.style[type])
   getTop() {
     return parseInt(this.element.style.top);
   }
